Add null driver to conditionally disable plugin
This plugin should be disabled for terminals and the git commit window.
I've added a "null class" to follow the same interface, but it doesn't
persist anything.

You can technically configure this through autocommands and setting
`b:deja_save_mode`, but that's a terrible experience. I want a better
way to override this while still providing defaults.